ReusableComponents/CustomModules/llm/requirements.txt
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.35
langchain_text_splitters==0.2.2
langchain_huggingface==0.0.3
langchain_anthropic==0.1.19
langchain-openai==0.1.23
python-dotenv==1.0.1